Found better way of dealing with Animations across threads

+++ src/Graphics/AnimationFlashing.cpp
@@ -1,6 +1,6 @@
 #include "AnimationFlashing.h"
 
-void AnimationFlashing::run(int32_t millis) {
+void AnimationFlashing::run(int64_t millis) {
 	if (millis > _lastTimeMillis + _interval) {
 		_lastTimeMillis = millis;
